fix IN token handling.
there was a bug in the way we configure IN endpoints, that was causing
the endpoint to send an initial empty packet to the host, and the driver
would keep sending data following the "transfer complete" signal of that
packet. this is wrong. the endpoint must be configured to NAK everything,
and when the user sets some data in PMA and marks the endpoint as VALID,
a transfer happens.

to make this work, we have to active poll the callbacks for new data, when
the endpoint is in NAK state. to avoid spamming this, we do it on SOF, so
we limit it to once per millisecond.